pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["TGRodriguez <tgrodriguez@fi.uba.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
black = "^23.12.1"
flakeheaven = "^3.3.0"
python-dotenv = "^1.0.0"

[tool.black]
line-length = 90

[tool.flakeheaven.plugins]
# include everything but W503: line break before binary operator
# because black does that :P
pyflakes = ["+*"]
pycodestyle = ["+*", "-W503"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
